requirements.txt
django=3
django-cors-headers
django-restframework